+++ lib/simple_parallax_web.dart
@@ -1,0 +1,8 @@
+import 'package:flutter_web_plugins/flutter_web_plugins.dart';
+/// SimpleParallax for web
+class SimpleParallaxWeb {
+  /// Register for web
+  static void registerWith(Registrar registrar) {
+    // Register the web-specific implementations here
+  }
+}