+++ lib/src/common/scroll_notifier.dart
@@ -1,15 +1,24 @@
 import 'package:flutter/cupertino.dart';
-
-/// Scroll notifier class
+/// Scroll notifier for parallax container
 class ScrollNotifier extends ChangeNotifier {
   double _scrollOffset = 0;
-
-  /// Get scrolling offset pixels
+  double _speed = 0.3;
+  /// Get actual scroll position
   double get scrollOffset => _scrollOffset;
-
-  /// Set scrolling offset pixels and notify it
+  /// Get parallax speed
+  double get speed => _speed;
+  /// Set scroll position
   void setScrollOffset(double offset) {
-    _scrollOffset = offset;
-    notifyListeners();
+    if (offset != _scrollOffset) {
+      _scrollOffset = offset;
+      notifyListeners();
+    }
+  }
+  /// Set parallax speed
+  void setSpeed(double speed) {
+    if (speed != _speed) {
+      _speed = speed;
+      notifyListeners();
+    }
   }
 }