+++ lib/src/components/parallax_layer.dart
@@ -1,9 +1,19 @@
 import 'package:flutter/material.dart';
 import 'package:provider/provider.dart';
-import 'package:simple_parallax/src/common/scroll_notifier.dart';
+import '../common/scroll_notifier.dart';
 
 /// Parallax layer for background image
 class ParallaxLayer extends StatelessWidget {
+  /// Layer constructor
+  const ParallaxLayer({
+    super.key,
+    required this.imagePath,
+    this.speed = 0.3,
+    this.decal = 1.0,
+    this.width,
+    this.height,
+  });
+
   /// Background image path
   final String imagePath;
 
@@ -19,23 +29,13 @@ class ParallaxLayer extends StatelessWidget {
   /// Background layer height
   final double? height;
 
-  /// Layer constructor
-  const ParallaxLayer({
-    super.key,
-    required this.imagePath,
-    this.speed = 0.3,
-    this.decal = 1.0,
-    this.width,
-    this.height,
-  });
-
   @override
   Widget build(BuildContext context) {
     return Consumer<ScrollNotifier>(
-      builder: (context, notifier, _) {
-        final blocHeight = height ?? MediaQuery.of(context).size.height;
-        final blocWidth = width ?? MediaQuery.of(context).size.width;
-        final imageHeight = blocHeight * (speed + decal);
+      builder: (BuildContext context, ScrollNotifier notifier, _) {
+        final double blocHeight = height ?? MediaQuery.of(context).size.height;
+        final double blocWidth = width ?? MediaQuery.of(context).size.width;
+        final double imageHeight = blocHeight * (speed + decal);
 
         double offset = notifier.scrollOffset;
 
@@ -43,14 +43,17 @@ class ParallaxLayer extends StatelessWidget {
           offset = 0;
         }
 
-        final imageActualHeight = imageHeight + offset;
+        final double imageActualHeight = imageHeight + offset;
 
         return Positioned(
           top: -offset,
           left: 0,
           width: blocWidth,
           height: imageActualHeight,
-          child: Image.asset(imagePath, fit: BoxFit.cover),
+          child: Image.asset(
+            imagePath,
+            fit: BoxFit.cover,
+          ),
         );
       },
     );