+++ CHANGELOG.md
@@ -1,9 +1,15 @@
 # SimpleParallax Versions
-## 0.0.1
 
-* Initial release
+## 0.1.0
+
+* Add parallax mods container and widget.
+* Rename all classes
 
 ## 0.0.2
 
 * Add comments
 * Update depedencies
+
+## 0.0.1
+
+* Initial release